inventoryManagement/requirements.txt
pymysql
flask==2.3.3
flask-cors
flask-mysql
python-dotenv